Feature #99: New email content filling system.
The mail content filling system is done. I applied it to the mail sent when a guest is created and to the mail sent when the password is reseted.

TODO: 1) apply it to the other mails 2) delete old unused mail templates 3) create the administrator view to change mail templates/subjects

To use it, you have to update your SQL schema and import the templates and subjects in your db : see Migration_v07_v08.sql added lines.